project/requirements.txt
alembic==1.6.2
aniso8601==9.0.1
bidict==0.21.2
blinker==1.4
click==7.1.2
dnspython==1.16.0
eventlet==0.31.0
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.14
Werkzeug==1.0.1